chore(glf): use shared feedback widget component
