+++ apps/go-local-first/package.json
@@ -100,11 +100,6 @@
 		"@tanstack/react-query": "5.54.1",
 		"@tanstack/react-query-devtools": "5.54.1",
 		"@tanstack/react-table": "^8.17.3",
-		"@tiptap/extension-highlight": "2.10.3",
-		"@tiptap/extension-link": "2.10.3",
-		"@tiptap/extension-typography": "2.10.3",
-		"@tiptap/react": "2.10.3",
-		"@tiptap/starter-kit": "2.10.3",
 		"@trpc/client": "next",
 		"@trpc/next": "next",
 		"@trpc/react-query": "next",
