+++ apps/go-local-first/src/app/_components/providers.tsx
@@ -1,13 +1,11 @@
 'use client'
 
 import * as React from 'react'
-import {
-	FeedbackProvider,
-	useFeedback,
-} from '@/feedback-widget/feedback-context'
 import { MuxPlayerProvider } from '@/hooks/use-mux-player'
 import { MDXProvider } from '@mdx-js/react'
 import { SessionProvider } from 'next-auth/react'
+
+import { FeedbackProvider } from '@coursebuilder/ui/feedback-widget/feedback-context'
 
 import AmplitudeContextProvider from './amplitude-provider'
