+++ apps/go-local-first/src/app/layout.tsx
@@ -6,12 +6,12 @@ import { Gabarito } from 'next/font/google'
 import { Party } from '@/app/_components/party'
 import { Providers } from '@/app/_components/providers'
 import Footer from '@/components/app/footer'
+import { FeedbackInsert } from '@/components/feedback-widget/feedback-insert'
 import Navigation from '@/components/navigation'
 import { ThemeProvider } from '@/components/theme-provider'
 import config from '@/config'
 import { courseBuilderAdapter } from '@/db'
 import { env } from '@/env.mjs'
-import { FeedbackInsert } from '@/feedback-widget/feedback-insert'
 import { getProduct } from '@/lib/products-query'
 import { TRPCReactProvider } from '@/trpc/react'
 import { ourFileRouter } from '@/uploadthing/core'
