+++ apps/go-local-first/src/components/feedback-widget/feedback-insert.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import * as React from 'react'
-import { useFeedback } from '@/feedback-widget/feedback-context'
+
+import { useFeedback } from '@coursebuilder/ui/feedback-widget'
 
 export function FeedbackInsert() {
 	const { isFeedbackDialogOpen, feedbackComponent } = useFeedback()
