+++ apps/go-local-first/src/components/navigation/index.tsx
@@ -4,7 +4,6 @@ import * as React from 'react'
 import Link from 'next/link'
 import { useParams, usePathname, useRouter } from 'next/navigation'
 import config from '@/config'
-import { useFeedback } from '@/feedback-widget/feedback-context'
 import { useSaleToastNotifier } from '@/hooks/use-sale-toast-notifier'
 import { api } from '@/trpc/react'
 import { cn } from '@/utils/cn'
@@ -12,6 +11,7 @@ import { X } from 'lucide-react'
 import { useSession } from 'next-auth/react'
 
 import { Button } from '@coursebuilder/ui'
+import { useFeedback } from '@coursebuilder/ui/feedback-widget/feedback-context'
 
 import { useNavLinks } from '../app/use-nav-links'
 import { LogoMark } from '../logo'
